FISH-6409 fixed mp-config-tck
Added BeforeDeploy observer to modify beans definition
