--- MicroProfile-Config/src/main/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,0 @@
-fish.payara.microprofile.config.tck.ArquillianExtension
