+++ MicroProfile-Config/src/test/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,0 +1,1 @@
+fish.payara.mptck.extension.ArquillianExtension
