+++ MicroProfile-Config/src/test/resources/beans.xml
